+++ Sources/EmojiKit/_Deprecations/Deprecations.swift
@@ -1,4 +1,5 @@
 import Foundation
+import SwiftUI
 
 public extension Emoji {
 
@@ -9,6 +10,9 @@ public extension Emoji {
     ) -> Bool {
         matches(query, in: locale)
     }
+    
+    @available(*, deprecated, renamed: "GridSectionTitleParameters")
+    typealias GridSectionParameters = GridSectionTitleParameters
 }
 
 extension EmojiCategory: CaseIterable {}
@@ -31,6 +35,68 @@ public extension EmojiCategory {
     
     @available(*, deprecated, renamed: "standard")
     static var standardWithoutFrequent: [EmojiCategory] { standard }
+}
+
+public extension EmojiGrid {
+    
+    @available(*, deprecated, message: "The emoji grid no longer takes a persisted category")
+    init(
+        axis: Axis.Set = .vertical,
+        emojis: [Emoji] = [],
+        categories: [EmojiCategory] = .standard,
+        query: String = "",
+        selection: Binding<Emoji.GridSelection> = .constant(.init()),
+        persistedCategory: EmojiCategory.PersistedCategory,
+        geometryProxy: GeometryProxy? = nil,
+        action: @escaping (Emoji) -> Void = { _ in },
+        categoryEmojis: @escaping (EmojiCategory) -> [Emoji] = { $0.emojis },
+        @ViewBuilder section: @escaping (Emoji.GridSectionParameters) -> SectionTitle,
+        @ViewBuilder item: @escaping (Emoji.GridItemParameters) -> GridItem
+    ) {
+        self.init(
+            axis: axis,
+            emojis: emojis,
+            categories: categories,
+            query: query,
+            selection: selection,
+            geometryProxy: geometryProxy,
+            action: action,
+            categoryEmojis: categoryEmojis,
+            section: section,
+            item: item
+        )
+    }
+}
+
+public extension EmojiScrollGrid {
+    
+    @available(*, deprecated, message: "The emoji grid no longer takes a persisted category")
+    init(
+        axis: Axis.Set = .vertical,
+        emojis: [Emoji] = [],
+        categories: [EmojiCategory] = .standard,
+        query: String = "",
+        selection: Binding<Emoji.GridSelection> = .constant(.init()),
+        persistedCategory: EmojiCategory.PersistedCategory,
+        geometryProxy: GeometryProxy? = nil,
+        action: @escaping (Emoji) -> Void = { _ in },
+        categoryEmojis: @escaping (EmojiCategory) -> [Emoji] = { $0.emojis },
+        @ViewBuilder section: @escaping (Emoji.GridSectionParameters) -> SectionTitle,
+        @ViewBuilder item: @escaping (Emoji.GridItemParameters) -> GridItem
+    ) {
+        self.init(
+            axis: axis,
+            emojis: emojis,
+            categories: categories,
+            query: query,
+            selection: selection,
+            geometryProxy: geometryProxy,
+            action: action,
+            categoryEmojis: categoryEmojis,
+            section: section,
+            item: item
+        )
+    }
 }
 
 public extension Collection where Element == Emoji {